pyproject.toml
[tool.poetry]
package-mode = false
name = "compliance-atom"
version = "1.0.0"
description = "Automatic compliance"
authors = [
    "Lukash Vladislav <v1r71k@yandex.ru>",
	"akkrevsky <@akkrevsky>",
	"vkireev1982",
    "holliwid",
    "inanbanan"
]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = ">=2.2,<3.0"
numpy = ">=1.26,<2.0"
pyarrow = ">=11.0,<12.0"
pydantic = ">=2.7,<3.0"
environs = "^11.0.0"
python-docx = "^1.1.2"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
pylint = "^3.2.5"
black = ">=23.9,<24.0"
flake8 = ">=6.1,<7.0"
flake8-bugbear = "^24.4.26"
flake8-pyproject = "^1.2.3"
isort = "^5.13.2"
pre-commit = "^3.7.1"
ipykernel = "^6.29.5"
jupyter = "^1.0.0"
pandas-stubs = "^2.2.2.240603"
matplotlib = "^3.9.1"
seaborn = "^0.13.2"
tqdm = "^4.66.5"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"

[tool.flake8]
max-line-length = 100
exclude = [
    ".git",
    "__pycache__",
    "*/__init__.py",
    "tests/*"
]
extend-ignore = [
    "E203",
    "E501",
    "W503",
    "B006",
    "B008",
    "B009",
    "B010"
]
select = [
    "C",
    "E",
    "F",
    "W",
    "B",
    "B950"
]

[tool.isort]
profile = "black"
skip = ["__init__.py"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
skip-string-normalization = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests/test.py",
    "tests/*_test.py",
    "tests/test_*.py",
]
